Conflicts:
	src/System.Security.Cryptography.Encoding/src/Configurations.props
	src/System.Security.Cryptography.Encoding/src/System.Security.Cryptography.Encoding.csproj

+++ cross/build-rootfs.sh
@@ -2,14 +2,21 @@
 
 usage()
 {
-    echo "Usage: $0 [BuildArch] [UbuntuCodeName]"
-    echo "BuildArch can be: arm, armel, arm64, x86"
-    echo "UbuntuCodeName - optional, Code name for Ubuntu, can be: trusty(default), vivid, wily, xenial. If BuildArch is armel, jessie(default) or tizen."
+    echo "Usage: $0 [BuildArch] [LinuxCodeName] [--skipunmount]"
+    echo "BuildArch can be: arm(default), armel, arm64, x86"
+    echo "LinuxCodeName - optional, Code name for Linux, can be: trusty(default), vivid, wily, xenial. If BuildArch is armel, LinuxCodeName is jessie(default) or tizen."
+    echo "--skipunmount - optional, will skip the unmount of rootfs folder."
     exit 1
 }
 
+__LinuxCodeName=trusty
 __CrossDir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 __InitialDir=$PWD
+__BuildArch=arm
+__UbuntuArch=armhf
+__UbuntuRepo="http://ports.ubuntu.com/"
+__LLDB_Package="lldb-3.6-dev"
+__SkipUnmount=0
 
 # base development support
 __UbuntuPackages="build-essential"
@@ -19,25 +26,15 @@ __UbuntuPackages+=" symlinks"
 
 # CoreCLR and CoreFX dependencies
 __UbuntuPackages+=" gettext"
-__UbuntuPackages+=" libcurl4-openssl-dev"
-__UbuntuPackages+=" libicu-dev"
-__UbuntuPackages+=" libkrb5-dev"
-__UbuntuPackages+=" liblttng-ust-dev"
-__UbuntuPackages+=" libssl-dev"
 __UbuntuPackages+=" libunwind8-dev"
+__UbuntuPackages+=" liblttng-ust-dev"
+__UbuntuPackages+=" libicu-dev"
+
+# CoreFX dependencies
+__UbuntuPackages+=" libcurl4-openssl-dev"
+__UbuntuPackages+=" libkrb5-dev"
+__UbuntuPackages+=" libssl-dev"
 __UbuntuPackages+=" zlib1g-dev"
-
-if [ -z "$LLVM_ARM_HOME" ]; then
-    __LLDB_Package="lldb-3.6-dev"
-fi
- 
-
-
-__BuildArch=arm
-__UbuntuArch=armhf
-__UbuntuCodeName=trusty
-__UbuntuRepo="http://ports.ubuntu.com/"
-__MachineTriple=arm-linux-gnueabihf
 
 __UnprocessedBuildArgs=
 for i in "$@" ; do
@@ -50,53 +47,53 @@ for i in "$@" ; do
         arm)
             __BuildArch=arm
             __UbuntuArch=armhf
-            __MachineTriple=arm-linux-gnueabihf
             ;;
         arm64)
             __BuildArch=arm64
             __UbuntuArch=arm64
-            __MachineTriple=aarch64-linux-gnu
-            ;;
-        x86)
-            __BuildArch=x86
-            __UbuntuArch=i386
-            __UbuntuRepo="http://archive.ubuntu.com/ubuntu"
             ;;
         armel)
             __BuildArch=armel
             __UbuntuArch=armel
             __UbuntuRepo="http://ftp.debian.org/debian/"
-            __MachineTriple=arm-linux-gnueabi
-            __UbuntuCodeName=jessie
+            __LinuxCodeName=jessie
+            ;;
+        x86)
+            __BuildArch=x86
+            __UbuntuArch=i386
+            __UbuntuRepo="http://archive.ubuntu.com/ubuntu/"
             ;;
         vivid)
-            if [ "$__UbuntuCodeName" != "jessie" ]; then
-                __UbuntuCodeName=vivid
+            if [ "$__LinuxCodeName" != "jessie" ]; then
+                __LinuxCodeName=vivid
             fi
             ;;
         wily)
-            if [ "$__UbuntuCodeName" != "jessie" ]; then
-                __UbuntuCodeName=wily
+            if [ "$__LinuxCodeName" != "jessie" ]; then
+                __LinuxCodeName=wily
             fi
             ;;
         xenial)
-            if [ "$__UbuntuCodeName" != "jessie" ]; then
-                __UbuntuCodeName=xenial
+            if [ "$__LinuxCodeName" != "jessie" ]; then
+                __LinuxCodeName=xenial
             fi
             ;;
         jessie)
-            __UbuntuCodeName=jessie
+            __LinuxCodeName=jessie
             __UbuntuRepo="http://ftp.debian.org/debian/"
             ;;
         tizen)
             if [ "$__BuildArch" != "armel" ]; then
-                echo "Tizen is available only for armel"
+                echo "Tizen is available only for armel."
                 usage;
                 exit 1;
             fi
-            __UbuntuCodeName=
+            __LinuxCodeName=
             __UbuntuRepo=
             __Tizen=tizen
+            ;;
+        --skipunmount)
+            __SkipUnmount=1
             ;;
         *)
             __UnprocessedBuildArgs="$__UnprocessedBuildArgs $i"
@@ -119,17 +116,25 @@ if [[ -n "$ROOTFS_DIR" ]]; then
     __RootfsDir=$ROOTFS_DIR
 fi
 
-umount $__RootfsDir/*
-rm -rf $__RootfsDir
+if [ -d "$__RootfsDir" ]; then
+    if [ $__SkipUnmount == 0 ]; then
+        umount $__RootfsDir/*
+    fi
+    rm -rf $__RootfsDir
+fi
 
-if [[ -n $__UbuntuCodeName ]]; then
-    qemu-debootstrap --arch $__UbuntuArch $__UbuntuCodeName $__RootfsDir $__UbuntuRepo
-    cp $__CrossDir/$__BuildArch/sources.list.$__UbuntuCodeName $__RootfsDir/etc/apt/sources.list
+
+if [[ -n $__LinuxCodeName ]]; then
+    qemu-debootstrap --arch $__UbuntuArch $__LinuxCodeName $__RootfsDir $__UbuntuRepo
+    cp $__CrossDir/$__BuildArch/sources.list.$__LinuxCodeName $__RootfsDir/etc/apt/sources.list
     chroot $__RootfsDir apt-get update
     chroot $__RootfsDir apt-get -f -y install
     chroot $__RootfsDir apt-get -y install $__UbuntuPackages
     chroot $__RootfsDir symlinks -cr /usr
-    umount $__RootfsDir/*
+
+    if [ $__SkipUnmount == 0 ]; then
+        umount $__RootfsDir/*
+    fi
 elif [ "$__Tizen" == "tizen" ]; then
     ROOTFS_DIR=$__RootfsDir $__CrossDir/$__BuildArch/tizen-build-rootfs.sh
 else
